requirements.txt
selenium>=2.53.1
django>=1.8
django-ckeditor>=5.0
django-compressor>=1.6
django-favicon>=0.1
mysqlclient>=1.3.7
htmlmin>=0.1.10
django-htmlmin>=0.8
Pillow>=3.0
djangorestframework>=3.3.3